regularize repeatmasker

+++ modules/nf-core/repeatmodeler/repeatmasker/main.nf
@@ -31,10 +31,11 @@ process REPEATMODELER_MASKER {
         $ref \\
         $args \\
 
-   mv genome.fasta.masked ${prefix}.masked
-   mv genome.fasta.out    ${prefix}.out
-   mv genome.fasta.tbl    ${prefix}.tbl
-   mv genome.fasta.cat    ${prefix}.cat
+   mv ${ref}.masked ${prefix}.masked
+   mv ${ref}.out    ${prefix}.out
+   mv ${ref}.tbl    ${prefix}.tbl
+   if [ -e ${ref}.cat ] ; then mv ${ref}.cat ${prefix}.cat; fi
+   if [ -e ${ref}.cat.gz ] ; then mv ${ref}.cat.gz ${prefix}.cat; fi
 
     cat <<-END_VERSIONS > versions.yml
     "${task.process}":
